Well hey! Welcome back! It's time to put our new object-oriented skills into practice. We're working on the same out of this world project: it has ships, you choose them, then they engage in epic battle!
In an editor, far far away, you'll see a simple application that runs this: index.php is the homepage and battle.php does the magic and shows the results. Last time, we created a single class called Ship, which describes all its properties - it's like a container for one ship's details:

 117 lines  lib/Ship.php
... lines 1 - 2
class Ship
{
    private {dollar_sign}name;
    private {dollar_sign}weaponPower = 0;
    private {dollar_sign}jediFactor = 0;
    private {dollar_sign}strength = 0;
    private {dollar_sign}underRepair;
... lines 14 - 115
}
We used this to replace these big associative arrays. Now we deal with cute Ship objects:

 126 lines  functions.php
... lines 1 - 4
function get_ships()
{
    {dollar_sign}ships = array();
    {dollar_sign}ship = new Ship('Jedi Starfighter');
    //{dollar_sign}ship->setName('Jedi Starfighter');
    {dollar_sign}ship->setWeaponPower(5);
    {dollar_sign}ship->setJediFactor(15);
    {dollar_sign}ship->setStrength(30);
    {dollar_sign}ships['starfighter'] = {dollar_sign}ship;
... lines 15 - 33
    return {dollar_sign}ships;
... lines 35 - 126
Remove all the Flat Functions!

Having a huge list of flat functions in functions.php is not a good recipe for staying organized. But in just a few minutes, we'll use some new classes to give our app a whole new level of sophistication. We'll get rid of battle() first.

Look at Ship: this is a class that basically just holds data - some people call that 'state', but I'll say 'data' - and I'm talking about the values on a Ship object's properties. So a Ship object holds data, but it doesn't really do any work. Sure, it has some methods on it, but these just return that data, after doing some small logic at best.

Reason #1 for creating a class is this: we need some organized unit to hold data.

But there's a second big reason to create a class: because you need to do some work. For example, in functions.php, the battle() function does work: we give it 2 Ships, it does some calculations, executes logic to see how different strengths affect each other and ultimately returns the result of that work.

And we're all familiar with creating functions like this. And here's the secret for OO: whenever you get the urge to create a flat function like battle(), don't. Instead, create a class and with a method inside of it.