Since this function is all about battling, let's create a new class called BattleManager. Be as creative as you want with naming: I want to describe that methods in this class will do things related to battling. Go copy and remove the flat battle() function: paste it into BattleManager. Put public in front of function. Remember, public means that code outside of this class will be able to call this. And yes, you don't have to add public: functions default to public if you say nothing, but let's keep things clear!
That's all you need to change: functions work the same inside or outside of a class: they have arguments, they return stuff.

But we do need to change code where we call this function - in battle.php. So how can we call this? Well, when we want to call a method on Ship, we need to have a Ship object first. The same is true here: we need a BattleManager object first. Start with a new variable called  and create a new BattleManager object.

At the top of functions.php, we have access to the Ship class because we're requiring it. Do the same for BattleManager.

There is a way where you can reference classes like BattleManager without needing to worry about the require statements. It's called autoloading, it's really common, and you'll learn how to master it in a future episode. But until then: if you have a class, require it.

Now we have 2 reasons to create a class. First, if you have some data - like properties that describe a ship, creating a class for that is nice. You'll create a Ship object whenever you have a set of that data. In get_ships(), we create 4 Ship objects. These types of classes are sometimes called models, because they model something, like a ship.

Second, if you need to make a function that does some work: create a class and put a method in it, like BattleManager. Or, you may put multiple methods inside one class - as long as they are all thematically similar.

You'll create one of these objects - like BattleManager - just one time, before you need to call a method on it. These are sometimes called service classes, because they perform work or service. Organizing your code to use service classes can be tricky, but we'll learn all about that.