didJediDestroyShipUsingTheForce function lived at the bottom of functions.php. In our app, this is only called from inside battle(), and since it obviously relates to battles, let's move it into BattleManager. Make it a private function.
Why did I make it private? Well, do we need use this function from outside of this class? No - the only code using it is up in battle(), so this is a perfect candidate to be private.

Above in battle(), update the calls to be {dollar_sign}this->didJediDestroyShipUsingTheForce(). The 'force' of our app is happy again.

Now, if someday we did want to use this function from outside of BattleManager, then we could change it to public. Ok, so why not just make everything public - isn't that more flexible? Yes, but making this private is nice: it means that if I want to change this function - add arguments or even change what it returns - I know that the only code that will be affected will be right inside this class. If it's public, who knows what code I might break in my app?

Start with private, make it public only if you need. The same rule goes for protected - something we'll talk about later with inheritance.

Let's make sure we didn't bust things. Refresh!