So let's use it inside battle(): instead of returning that array, return a new BattleResult and pass it {dollar_sign}usedJediPowers, {dollar_sign}winningShip and {dollar_sign}losingShip.
But hey, we're referencing a class, so make sure you require it in bootstrap.php.

So where is battle() being called? It's at the top of battle.php - and this {dollar_sign}outcome variable used to be that associative array - now it's a fancy BattleResult object:

 99 lines  battle.php
... lines 1 - 30
{dollar_sign}outcome = {dollar_sign}battleManager->battle({dollar_sign}ship1, {dollar_sign}ship1Quantity, {dollar_sign}ship2, {dollar_sign}ship2Quantity);
... lines 32 - 99
This means that our code below - the stuff that treats {dollar_sign}outcome like an array - should blow up.:

 99 lines  battle.php
... lines 1 - 70
  <?php if ({dollar_sign}outcome['winning_ship']): ?>
    <?php echo {dollar_sign}outcome['winning_ship']->getName(); ?>
  <?php else: ?>
    Nobody
  <?php endif; ?>
... lines 76 - 99
Let's see some fireworks! Boom error!

Cannot use object of type BattleResult as array on line 71.
But we do need to get the winning ship from the BattleResult object. Is that possible right now? No - the {dollar_sign}winningShip property is private. If we want to access it from outside the class, we need a public function that returns it for us. We did this same thing in Ship with methods like getName().