Now we can finally go back to battle.php and start using these public methods. Start by renaming {dollar_sign}outcome to {dollar_sign}battleResult - it's more clear this is a BattleResult object.
Below, use {dollar_sign}battleResult->getWinningShip().

Except, where's my auto-completion on that method? This will work, but PhpStorm is highlighting the method like it's wrong. It doesn't know that  is a BattleResult object.

Why? Look at battle(). We are returning a BattleResult, but oh no, the @return above this method still advertises that this method returns an array. Fix that with @return BattleResult.

Ok, now PhpStorm is acting friendly - the angry highightling on the method is gone. Now update the other spots: {dollar_sign}battleResult->getWinningShip()->getName(): thank you auto-complete. Use that same method once more, and in the if statement, use that nice wereJediPowersUsed() method. Finish with {dollar_sign}battleResult->getLosingShip()

I think we're done. Refresh to try it! Ship it!

And gone are the days of needing to use weird associative arrays: BattleManager::battle() returns a nice BattleResult object. And we're in full control of what public methods we put on that.