I need to show you something - so start another battle between some Jedi Star Fighters. It works... but if I refresh enough times... come on... yes! It blows up!
Argument 2 passed to BattleResult::__construct() must be an instance of Ship, null given.

In BattleResult - because we're good programmers - we type-hinted the two Ship arguments. Buuuuut, if you look at the battle() function, there's a case where the ships can destroy each other. And when that happens, there is no winning or losing ship - they're both null. Since - news flash null is not a Ship object, PHP gets angry and casts down this big error.

When you type-hint an argument, the value must be that class - not even null is ok. But sometimes you do have a spot where an argument might be a specific object, or it might be null. To support this, make the argument optional - add an = null after it.

I don't have to, but I'll update @return on the methods to be Ship|null.

PhpStorm will still give me auto-completion - but this is a signal to other developers not to blindly call this method and always assume it will return a Ship object. We're already coding safely in battle.php: we check to make sure getWinningShip() returns something before calling a method on it.