The takeaway is that if you change an object, you're changing that object everywhere. To prove this, take our {dollar_sign}ship1 and {dollar_sign}ship2 - which are not returned by the battle() function - and add a new section that prints the finished strength. Add a dl element to make them a little pretty.
We're missing auto-complete because we have some bad PHPDoc somewhere. We'll fix that in a bit.

Time to try it! Since objects are passed by reference, we should see the new, modified strength values - not the originals. Absolutely perfect.

Now let's get really wild and start fetching our ships from a database.