Getting our Ship objects is easy: create a ShipLoader and call getShips() on it. We don't care how ShipLoader is getting these - that's its problem.
Hardcoding is so 1990, let's load objects from the database! We need to get these ships to their battlestations!

Database Setup

At the root of your project, open up a resources directory. Copy init_db.php out of there to the root of your project and open it up:
<?php
/*
 * SETTINGS!
 */
{dollar_sign}databaseName = 'oo_battle';
{dollar_sign}databaseUser = 'root';
{dollar_sign}databasePassword = '';
/*
 * CREATE THE DATABASE
 */
{dollar_sign}pdoDatabase = new PDO('mysql:host=localhost', {dollar_sign}databaseUser, {dollar_sign}databasePassword);
{dollar_sign}pdoDatabase->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
{dollar_sign}pdoDatabase->exec('CREATE DATABASE IF NOT EXISTS oo_battle');

This script will create a database and add a ship table with columns for id, name, weapon_power, jedi_factor, strength and is_under_repair:

 57 lines  init_db.php
... lines 1 - 25
->exec('CREATE TABLE 'ship' (
 'id' int(11) NOT NULL AUTO_INCREMENT,
 'name' varchar(255) COLLATE utf8mb4_unicode_ci NOT NULL,
 'weapon_power' int(4) NOT NULL,
 'jedi_factor' int(4) NOT NULL,
  int(4) NOT NULL,
 'is_under_repair' tinyint(1) NOT NULL,
 PRIMARY KEY ('id')
) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci');
... lines 35 - 57
At the bottom, it inserts 4 rows into that table for the 4 ships we have hardcoded right now:

 57 lines  init_db.php
... lines 1 - 38
{dollar_sign}pdo->exec('INSERT INTO ship
    (name, weapon_power, jedi_factor, strength, is_under_repair) VALUES
    ('Jedi Starfighter', 5, 15, 30, 0)'
);
{dollar_sign}pdo->exec('INSERT INTO ship
    (name, weapon_power, jedi_factor, strength, is_under_repair) VALUES
    ('CloakShape Fighter', 2, 2, 70, 0)'
);
{dollar_sign}pdo->exec('INSERT INTO ship
    (name, weapon_power, jedi_factor, strength, is_under_repair) VALUES
    ('Super Star Destroyer', 70, 0, 500, 0)'
);
{dollar_sign}pdo->exec('INSERT INTO ship
    (name, weapon_power, jedi_factor, strength, is_under_repair) VALUES
    ('RZ-1 A-wing interceptor', 4, 4, 50, 0)'
);
... lines 55 - 57
If we run this file, it should get everything powered up. Head to your browser and run it there:

http://localhost:8000/init_db.php
If you see - Ding! - you know it worked. If you see a terrible error, check the database credentials at the top - make sure the user can create a new database.

If you want to check the database with something like phpMyAdmin, you'll see one ship table with 4 rows.