You look ready to query, copy the two lines that create the PDO object in init_db and head into ShipLoader. Keep things simple: getShips() needs to make a query. So for now, paste the PDO lines right here. Update the database name to be oo_battle and I'll fill in root as the user.
Ok, query time! Create a  variable and set it to ->prepare() with the query inside - SELECT * FROM ship.

If PDO or prepared statements are new to you, don't worry - they're pretty easy. And besides, using PDO is another chance to play with objects!

Run {dollar_sign}statement->execute() to send the query into hyperdrive and create a new {dollar_sign}shipsArray that's set to {dollar_sign}statement->fetchAll() with an argument: PDO::FETCH_ASSOC. var_dump this variable.

This queries for every row and returns an associative array. The PDO::FETCH_ASSOC part is a class constant - a nice little feature of classes we'll talk about later.

Let's see what this looks like! Head back to the homepage and refresh!
Ok! 4 rows of data