Of course, what we need are objects, not arrays. But first, a quick piece of organization. Copy all this good PDO stuff and at the bottom, create a new private function queryForShips(). Paste here and return that .
Head back up, call this method, then remove the original code.

Make sure things still work - cool! Now, why did we do this? Well, we had a chunk of code that did something - it made a query. Moving it into its own function has two advantages. First, we can re-use it later if we need to. But more importantly, it gives the code a name: queryForShips(). Now it's easy to see what it does - a lot easier than when this was stuck in the middle of other code.

So, creating private functions to help split code into small chunks is awesome.