Ships are loading dynamically, buuuuuut, I've got some bad news: we broke our app. Start a battle - select the Jedi Starfighter as one of the ships and engage.
Huh, so instead of the results, we see:
Don't forget to select some ships to battle!

Pretty sure we selected a ship... But the URL has a ?error=missing_data part, index.php is reading this. It all comes from battle.php and it happens if we POST here, but we are missing ship1_name or ship2_name. In other words, if we forget to select a ship. But we did select a ship! Somehow, these select menus are broken. Check out the code: we're looping over  and using  as the option value:

 119 lines  index.php
... lines 1 - 90
  <select class='center-block form-control btn drp-dwn-width btn-default btn-lg dropdown-toggle' name='ship1_name'>
... line 92
    <?php foreach ({dollar_sign}ships as {dollar_sign}key => {dollar_sign}ship): ?>
      <?php if ({dollar_sign}ship->isFunctional()): ?>
        <option value='<?php echo {dollar_sign}key; ?>'><?php echo {dollar_sign}ship->getNameAndSpecs(); ?></option>
      <?php endif; ?>
    <?php endforeach; ?>
  </select>
... lines 99 - 119
In getShips(), the key was a nice, unique string. But now it's just the auto-increment index. The page fails because the 0 index looks like an empty string in battle.php.