We still need something unique so that we can tell battle.php exactly which ships are fighting. Fortunately, the ship table has exactly that: an auto-incrementing primary key id column. If we use this as the option value, we can query for the ships using that in battle.php. Blast off! I mean, we should totally do that.
In ShipLoader, we could put the id as the key of the array. But instead, since id is a column on the ship table, why not also make it a property on the Ship class? Open up Ship and add a new private {dollar_sign}id.

And at the bottom, right click, then make the getter and setter for the id property. Update the PHPDoc to show that {dollar_sign}id is an integer. Optional, but nice.

Now when we get our Ship objects, we need to call setId() to populate that property: {dollar_sign}ship->setId() and {dollar_sign}shipData['id']