Head over to index.php to use the fancy new property. Remove the {dollar_sign}key in the foreach - no need for that. And instead of the key, print {dollar_sign}ship->getId(). Also change the select name to be ship1_id so we don't get confused about what this value is.
Make the same changes in the second select: update the select name, remove {dollar_sign}key from the loop, and finish with {dollar_sign}ship->getId().

Ok, before we touch battle, try this out. No errors! And the select items have values 1, 2, 3 and 4 - the auto-increment ids in the database. Success!