We've renamed the select fields and we're sending a database id. Let's update battle.php for this. First, we need to change the  keys: look for ship1_id and ship2_id. Update the variables names too - {dollar_sign}ship1Id and {dollar_sign}ship2Id. That'll help us not get confused. Update the variables in the first if statement.
Before, we got all the  then used the array key to find the right ones. That won't work anymore - the key is just an index, but we have the id from the database.

Instead, we can use that id to query for a single ship's data. Where should that logic live? In ShipLoader! It's only job is to query for ship information, so it's perfect.

Create a new public function findOneById() with an {dollar_sign}id argument. Copy all the query logic from queryForShips() and put it here. For now don't worry about all this ugly code duplication. Update the query to be SELECT * FROM ship WHERE id = :id and pass that value to execute() with an array of id to {dollar_sign}id.

If this looks weird to you - it's a prepared statement. It runs a normal query, but prevents SQL injection attacks. Change the variable below to be  and change fetchAll() to just fetch() to return the one row. Dump this at the bottom.

Ok, back to battle.php! Let's use this. Now, {dollar_sign}ship1 = {dollar_sign}shipLoader->findOneById({dollar_sign}ship1Id). And {dollar_sign}ship2 = {dollar_sign}shipLoader->findOneById({dollar_sign}ship2Id). And I need to move this code further up above the bad_ships error message. We'll use it in a second.

Try it! Fight some Starfighters against a Cloakshape Fighter. There's the dump for just one row! Sweet, let's finish this!