The last step is to take this array and turn it into a Ship object. And good news! We've already done this in getShips()! And instead of repeating ourselves, this is another perfect spot for a private function. Create one called createShipFromData with an array {dollar_sign}shipData argument.
Copy all the new Ship() code and paste it here. Return the {dollar_sign}ship variable.

Now, anyone inside ShipLoader can call this, pass an array from the database, and get back a fancy new Ship object.

Back in getShips(), remove all that code and just use {dollar_sign}this->createShipFromData(). Do the same thing in findOneById().

In battle.php, {dollar_sign}ship1 and {dollar_sign}ship2 should now be Ship objects. The next if statement is a way to make sure that valid ship ids were passed: maybe someone is messing with our form! With these tough ships in my database I should hope not.

And that should do it!

Go back and load the homepage fresh. And start a battle. When we submit, we'll be POST'ing these 2 ids to battle.php. And it works!