Let's fix one little thing that's bothering me. In index.php, we call getShips(). But when we loop over , PhpStorm acts like all of the methods on the Ship object don't exist: getName not found in class.
If you look above getShips(), there's no PHP documentation. And so PhpStorm has no idea what this function returns. To fix that, add the /** above it and hit enter to generate some basic docs. Now it says @return array. That's true, but it doesn't tell it what's inside the array. Change it to @return Ship[].

This says: 'I return an array of Ship objects'. And when we loop over something returned by getShips(), we get happy code completion. Do the same thing above findOneById() - it returns just one Ship or null