Ok, next problem: at the bottom of ShipLoader, our database connection information is hardcoded. That's a problem for two reasons. First, if this works on my computer, it probably won't work on production, unless everything matches up. And second, what if we need a database connection inside some other class? Right now, we'd just have to copy and paste those credentials into yet another spot. Eww.
Here's the goal: move the database configuration out of this class to somewhere more central so it can be re-used. And good news: the way you do this is fundamentally important to using object-oriented code correctly.

How to Make the OO Kittens Sad

But first, let me tell you what you shouldn't do. You shouldn't just move this configuration to another file and then use some global keywords to get that information here. You will see this kind of stuff - heck you might see it all the time depending on your project. The problem is that your code gets harder to read and maintain: 'Hey, where the heck is this {dollar_sign}dbPassword' variable created? And what if you wanted to re-use this class in another project? It better have global variables with the exact same names.

Learning the better way is the difference between an 'ok' object-oriented developer and a great one: and even though this is only episode 2, you're about to learn it.

The Secret: Pass Objects the Config they Need

The secret is this: if a service class - like ShipLoader - needs information - like a database password - we need to pass that information to ShipLoader instead of expecting it to use a global keyword or some other method to 'find' it on its own. The most common way to do this is by creating a constructor.