Create a public function __construct() and make an argument for each piece of configuration this class needs. ShipLoader needs three pieces of configuration. First, the database DSN - which is the connection parameter, thing mysql:host=localhost. It also needs the {dollar_sign}dbUser and the {dollar_sign}dbPassword.
And just like any class, you'll set each of these on a private property. Create a private {dollar_sign}dbDsn, {dollar_sign}dbUser and {dollar_sign}dbPass. In __construct(), assign each argument to the property. I made my arguments - like {dollar_sign}dbUser the same as my property name - but that's not needed, it's just nice for my own sanity.

If this feels silly, pointless or you don't get it yet. That's GREAT. Keep watching. Thanks to this change, whoever creates a new ShipLoader() is forced to pass in these 3 configuration arguments. We don't care who creates ShipLoader, but when they do, we store the configuration on three properties and can use that stuff in our methods below.

At the bottom - let's do that. Copy the long database DSN string from new PDO() and replace it with {dollar_sign}this->dbDsn. Make the second argument {dollar_sign}this->dbUser and the third {dollar_sign}this->dbPass

And this class is done!