But now, when we create ShipLoader, we need to pass arguments. In index.php, PhpStorm is angry - required parameter {dollar_sign}dbDsn - we're missing the first argument. We could just paste our database credentials right here. But we'll probably want them somewhere central.
Open bootstrap.php and create a new {dollar_sign}configuration array. We'll use this now as sort of a 'global configuration' variable. Put the 3 database credential things here .

Since we're requiring this from index.php, we can just use it there: {dollar_sign}configuration['db_dsn] is the first argument then use db_user as the second argument and db_pass to finish things off.

Yes! Now the app's configuration is all in one file. In index.php, we pass this stuff to ShipLoader via its __construct() method. Then ShipLoader doesn't have any hardcoded configuration. Anything that was hardcoded before was replaced by a __construct() argument and a private property.

Make sure our ships are still battling. Refresh! Still not broken!

The Big Important Rule

Here's the rule to remember: don't put configuration inside of a service class. Replace that hardcoded configuration with an argument. This allows anyone using your class to pass in whatever they want. The hardcoding is gone, and your class is more flexible.

Oh, and by the way - this little strategy is called dependency injection. Scary! It's a tough concept for a lot of people to understand. If it's not sinking in yet, don't worry. Practice makes perfect.