Ready for the next problem? Our PDO object is configurable, but we're still creating it inside of ShipLoader. What's going to happen if we add a battle table and a BattleLoader? Will it also need to create its own PDO object? Right now - yea. So if we have 50 tables, that means 50 separate connections. The horror!
I want one connection that every class uses.

Here's the goal: move the new PDO() call out of ShipLoader so that it can be created in a central location and used by everyone. How? By using the same strategy we just learned with configuration. If you want to move something out of a service class, add it as a __construct() argument and pass it in.