Let's do it! Instead of passing in the 3 database options, we need to pass in the whole PDO object. Replace the 3 arguments with just one: {dollar_sign}pdo. Give it a type-hint to be great programmers. Next, remove the three configuration properties. And back in __construct(), we already have a {dollar_sign}pdo property, so set that with {dollar_sign}this->pdo = {dollar_sign}pdo.
Time to simplify the getPDO() function. We don't need to worry about creating the object anymore. Instead, just return the property.

Again: big picture: if you need to remove something from a service class - whether it's configuration or an object - remove it, and add it as an argument to the __construct() function.