But now, we need go to index.php and change the arguments we're passing to the new ShipLoader(). We're not passing these three configuration pieces anymore. Copy those. Above this, create the PDO object. {dollar_sign}pdo = new PDO() and paste in the arguments.
Pass {dollar_sign}pdo as the only argument to new ShipLoader().
Ok, let's try it! Still works. Geez - we're unstoppable today.

Unfortunately, this isn't the only place we need this. Copy the  and {dollar_sign}shipLoader code and paste it into battle.php.
Choose some ships to battle and.... Engage. And that still works too!

The Big Important Takeaway

Ready for the big important takeaway? Don't include configuration or create new service objects from within a service. Even though the PDO class comes from PHP, it is a service class: it does work. If we create that service object from within a class, we can't easily share it or control it.

Instead, create all of your service objects in one place and then pass them into each other. This stuff is hard - a lot of systems violate the heck out of these rules! And that's ok - I want you to learn to become a great object-oriented developer, so we're looking at the best way to do things.

The downside is that the code to create the service objects is getting a bit complicated. And it's duplicated! Dang it - it's not right yet. Let's fix that next by learning another awesome strategy.