Ok, we've started removing duplication. But I made us go one step backwards: once again, our configuration is buried inside a class - I'd rather have that somewhere central. Fix this like we always do when we want to remove some details from a class: create a public function __construct() with a configuration argument. Add the configuration property and assign it in the construct function.
Down in getPDO(), let's celebrate! Remove the  variable and reference the property instead.

This is an easy change - bootstrap.php already holds the central {dollar_sign}configuration array. In battle.php pass {dollar_sign}configuration to the Container.
And do the same thing for index.php.

This keeps my requirement of a centralized configuration array and centralizing where we create service objects. But we still need to move a few more service objects in here and fix one more issue. Almost there!