Keep going! We don't want to instantiate a ShipLoader object manually in battle.php and index.php. Let's just do it inside Container.
Follow the same pattern: create a private property called shipLoader, and a public function getShipLoader().

In here, add the same if statement: if ({dollar_sign}this->shipLoader === null), then {dollar_sign}this->shipLoader = new ShipLoader(). Remember, it has a required argument for the PDO object. That's easy, just say {dollar_sign}this->getPDO(). At the bottom return {dollar_sign}this->shipLoader and add the PHPDoc above it.

Use it! In index.php, say {dollar_sign}shipLoader = {dollar_sign}container->getShipLoader(). And I have a bonus for you! We don't need the  variable anymore - we only did that to pass it to ShipLoader. Simplify!
Copy the new  line and repeat this in battle.php.
Ok, now it works.