Best Practices vs the Real World
In this episode, instead of learning more OO concepts, we went straight to the hard stuff and learned how to organize our code into model classes that hold data and service classes that do work. We also learned that when you're in a service class - like ShipLoader - instead of hardcoding configuration or creating other service objects inside, we can move those outside of the class and add anything we need as an argument to the __construct() function. Then, we'll pass that information to the class. That's dependency injection, and it's one of the harder things to grasp about OO. So if it doesn't totally make sense yet - stick with us - we'll keep practicing.

Now a quick warning. When you look at other projects, this idea of model objects -- that hold data but don't do anything - and service objects - that do work but don't really hold any data - is not always followed. Sometimes you'll see these mixed together you might have a class like Ship that has methods in it that do work - like battle() or even save() that would save the Ship's data to the database.

What I'm showing you are 'best practices'. When you get out into the wild, it's not always this clean. And that's ok - over time, you'll learn to bend the rules when it makes sense. But in your mind, keep these two types of classes separate and recognize if a class is a model, a service or both.

Ok guys - in the next episodes, we're going to dive into more great concepts of OO - like interfaces, abstract classes, and static calls. These will really take your mad-skills to the next level.